[Refactor] Extract MainView class in MyToolWindowFactory
- I think ToolWindowFactory doesn't have any responsbility about designing main view.
- So, I had extracted this mainView class.
- MyTestListPanel is always created by MyMainView. so, in my opinion, protected is more appropriate than public.
